// dfp-piao-fe/src/app/shared/models/classes/anagrafica-dto.ts
export class AnagraficaDTO {
  id?: number;
  idSezione1?: number;
  denominazioneEnte?: string;
  acronimoPA?: string;
  codiceFiscale?: string;
  codiceIPA?: string;
  tipologiaPA?: string;
  tipologiaIstat?: string;
  piva?: string;
  indirizzoSedeLegale?: string;
  indirizzoURP?: string;
  www?: string;
  mail?: string;
  telefono?: string;
  pec?: string;
  nomeRPCT?: string;
  cognomeRCTP?: string;
  ruoloRPCT?: string;
  dataNominaRPCT?: Date;
  nomeRTD?: string;
  strutturaRifRTD?: string;
  social?: string;
}